Make git 2.55.0 work
Carry forward the parse-options defval fix from git-2.46.0 (2c2f4ea90110
"Make git work") to git 2.55.0.

Fil-C pointers lose their capability when round-tripped through intptr_t,
so storing string or pointer default values in 'intptr_t defval' compiles
but crashes at runtime ("cannot read pointer with null object") as soon
as the default value is used, e.g. 'git status -u' or
'git archive --add-file'. Change struct option's defval to void *: integer
option macros cast their defval to (void *), string/pointer defvals are
stored as real pointers, and integer use sites in parse-options.c,
builtin/am.c and builtin/config.c cast back to intptr_t.

Compared to the 2.46.0 patch this is adapted to upstream's switch to
designated initializers for option tables and to the parse-options.c
rewrite (intmax_t/precision-based get_int_value/set_int_value and the new
signed_int_fits defval validation). The compat/bswap.h hunk of the old
patch is dropped: Fil-C's clang now compiles the x86 inline-asm bswap
correctly (verified with sha256 repos, commit-graph and multi-pack-index
read/write/verify).

+++ projects/git-2.55.0/parse-options.c
@@ -150,9 +150,9 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 	{
 		intmax_t value = get_int_value(opt, flags);
 		if (unset)
-			value &= ~opt->defval;
+			value &= ~(intptr_t)opt->defval;
 		else
-			value |= opt->defval;
+			value |= (intptr_t)opt->defval;
 		return set_int_value(opt, flags, value);
 	}
 
@@ -160,9 +160,9 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 	{
 		intmax_t value = get_int_value(opt, flags);
 		if (unset)
-			value |= opt->defval;
+			value |= (intptr_t)opt->defval;
 		else
-			value &= ~opt->defval;
+			value &= ~(intptr_t)opt->defval;
 		return set_int_value(opt, flags, value);
 	}
 
@@ -172,7 +172,7 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 		if (unset)
 			BUG("BITOP can't have unset form");
 		value &= ~opt->extra;
-		value |= opt->defval;
+		value |= (intptr_t)opt->defval;
 		return set_int_value(opt, flags, value);
 	}
 
@@ -195,7 +195,7 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 	}
 
 	case OPTION_SET_INT:
-		return set_int_value(opt, flags, unset ? 0 : opt->defval);
+		return set_int_value(opt, flags, unset ? 0 : (intptr_t)opt->defval);
 
 	case OPTION_STRING:
 		if (unset)
@@ -266,7 +266,7 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 		if (unset) {
 			value = 0;
 		} else if (opt->flags & PARSE_OPT_OPTARG && !p->opt) {
-			value = opt->defval;
+			value = (intptr_t)opt->defval;
 		} else if (get_arg(p, opt, flags, &arg)) {
 			return -1;
 		} else if (!*arg) {
@@ -295,7 +295,7 @@ static enum parse_opt_result do_get_value(struct parse_opt_ctx_t *p,
 		if (unset) {
 			value = 0;
 		} else if (opt->flags & PARSE_OPT_OPTARG && !p->opt) {
-			value = opt->defval;
+			value = (intptr_t)opt->defval;
 		} else if (get_arg(p, opt, flags, &arg)) {
 			return -1;
 		} else if (!*arg) {
@@ -678,7 +678,7 @@ static void parse_options_check(const struct option *opts)
 		case OPTION_NEGBIT:
 		case OPTION_BITOP:
 		case OPTION_COUNTUP:
-			if (!signed_int_fits(opts->defval, opts->precision))
+			if (!signed_int_fits((intptr_t)opts->defval, opts->precision))
 				optbug(opts, "has invalid defval");
 			/* fallthru */
 		case OPTION_NUMBER:
